+++ src/filters/biquadcombo.rs
@@ -114,40 +114,6 @@ impl BiquadCombo {
         filters
     }
 
-    fn make_peq5(
-        samplerate: usize,
-        f_all: [f64; 5],
-        q_all: [f64; 5],
-        g_all: [f64; 5],
-    ) -> Vec<biquad::Biquad> {
-        let mut filters = Vec::with_capacity(5);
-        for (n, ((f, q), g)) in f_all.iter().zip(q_all).zip(g_all).enumerate() {
-            if q.abs() > 0.001 {
-                let filtconf = match n {
-                    0 => config::BiquadParameters::Lowshelf(config::ShelfSteepness::Q {
-                        freq: *f,
-                        q,
-                        gain: g,
-                    }),
-                    4 => config::BiquadParameters::Highshelf(config::ShelfSteepness::Q {
-                        freq: *f,
-                        q,
-                        gain: g,
-                    }),
-                    _ => config::BiquadParameters::Peaking(config::PeakingWidth::Q {
-                        freq: *f,
-                        q,
-                        gain: g,
-                    }),
-                };
-                let coeffs = biquad::BiquadCoefficients::from_config(samplerate, filtconf);
-                let filt = biquad::Biquad::new("", samplerate, coeffs);
-                filters.push(filt);
-            }
-        }
-        filters
-    }
-
     fn make_graphic(
         samplerate: usize,
         freq_min: f32,
@@ -222,35 +188,6 @@ impl BiquadCombo {
             }
             config::BiquadComboParameters::Tilt { gain } => {
                 let filters = BiquadCombo::make_tilt(samplerate, gain);
-                BiquadCombo {
-                    samplerate,
-                    name,
-                    filters,
-                }
-            }
-            config::BiquadComboParameters::FivePointPeq {
-                fls,
-                qls,
-                gls,
-                fp1,
-                qp1,
-                gp1,
-                fp2,
-                qp2,
-                gp2,
-                fp3,
-                qp3,
-                gp3,
-                fhs,
-                qhs,
-                ghs,
-            } => {
-                let filters = BiquadCombo::make_peq5(
-                    samplerate,
-                    [fls, fp1, fp2, fp3, fhs],
-                    [qls, qp1, qp2, qp3, qhs],
-                    [gls, gp1, gp2, gp3, ghs],
-                );
                 BiquadCombo {
                     samplerate,
                     name,
@@ -336,31 +273,6 @@ pub fn validate_config(samplerate: usize, conf: &config::BiquadComboParameters) 
                 return Err(config::ConfigError::new("Gain must be > -100").into());
             } else if *gain >= 100.0 {
                 return Err(config::ConfigError::new("Gain must be < 100").into());
-            }
-            Ok(())
-        }
-        config::BiquadComboParameters::FivePointPeq {
-            fls,
-            qls,
-            fp1,
-            qp1,
-            fp2,
-            qp2,
-            fp3,
-            qp3,
-            fhs,
-            qhs,
-            ..
-        } => {
-            if *qls <= 0.0 || *qhs <= 0.0 || *qp1 <= 0.0 || *qp2 <= 0.0 || *qp3 <= 0.0 {
-                return Err(config::ConfigError::new("All Q-values must be > 0").into());
-            } else if *fls >= maxfreq
-                || *fhs >= maxfreq
-                || *fp1 >= maxfreq
-                || *fp2 >= maxfreq
-                || *fp3 >= maxfreq
-            {
-                return Err(config::ConfigError::new("All frequencies must be > 0").into());
             }
             Ok(())
         }